backendraef/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
gtts==2.5.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy<2.0
